refactor: [Lv.6] 6단계를 위한 2차 리팩토링

+++ src/main/java/org/example/expert/domain/comment/service/CommentAdminService.java
@@ -13,6 +13,9 @@ public class CommentAdminService {
 
     @Transactional
     public void deleteComment(long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new IllegalArgumentException("댓글이 존재하지 않습니다.");
+        }
         commentRepository.deleteById(commentId);
     }
 }
